+++ src/univariate/continuous/triweight.jl
@@ -14,7 +14,7 @@ end
 
 Triweight(μ::T, σ::T, ::NoArgCheck) where {T<:Real} = Triweight{T}(μ, σ)
 Triweight(μ::Real, σ::Real) = Triweight(promote(μ, σ)...)
-Triweight(μ::Integer, σ::Integer) = Triweight(Float64(μ), Float64(σ))
+Triweight(μ::Integer, σ::Integer) = Triweight(float(μ), float(σ))
 Triweight(μ::T) where {T <: Real} = Triweight(μ, one(T))
 Triweight() = Triweight(0.0, 1.0, NoArgCheck())
